Move shared display types to dedicated file
